schema: Add full support for OVSDB types
The way the column types are specified in RFC7047 is not directly
unmarshallable by the json encoder.

This patch adds the necessary structs and functions to manually
unmarshal the schema definition.

The result is the possibility of determining the exact time of golang
structure needed to store each column type.

Signed-off-by: Adrian Moreno <amorenoz@redhat.com>