row: Add native row and ovsRow to row conversion
Currently, Row is being used to store the result of a query or an input
to one. This struct contains ovs-specific objects such as UUID, OvsSet
and OvsMap.

As preliminary work to hide this API from the user, add a conversion
mechanism between OvsRows(containing ovs structs) and Rows (native objects).

Since the table schema is required for this, the actual implementation
is performed in the ColumnSchema struct

Signed-off-by: Adrian Moreno <amorenoz@redhat.com>